backend/requirements.txt
numpy
opencv-python-headless
mediapipe
Pillow
flask
flask_cors
gunicorn
requests